ibit/requirements.txt
imageio
imageio-ffmpeg
hydra-core
pandas
ipdb
tqdm
torchvision
yapf
mujoco_py
sklearn
matplotlib
kornia
scikit-image